requirements.txt
rembg
fastapi
uvicorn
Pillow
Jinja2
onnxruntime
python-multipart